add anthropic and azure, refactor llm factory
