+++ pyproject.toml
@@ -20,6 +20,7 @@ dependencies = [
   "importlib-resources>=6.4.5",
   "langchain>=0.3.3",
   "langchain_openai>=0.2.2",
+  "langchain-anthropic>=0.3.15",
   "langchain_aws>=0.2.2",
   "pydantic>=2.9.2",
   "hydra-core>=1.3",
